Took in comments
