+++ src/Build.UnitTests/BackEnd/ProcessIdTask.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Microsoft.Build.Framework;
 using Microsoft.Build.Utilities;
 using System;
 using System.Diagnostics;
@@ -11,13 +12,16 @@ namespace Microsoft.Build.UnitTests
     /// </summary>
     public class ProcessIdTask : Task
     {
+        [Output]
+        public Int32 pid { get; set; }
+
         /// <summary>
         /// Log the id for this process.
         /// </summary>
         /// <returns></returns>
         public override bool Execute()
         {
-            Log.LogMessage("PID to shut down is " + Process.GetCurrentProcess().Id + " (EndPID)");
+            pid = Process.GetCurrentProcess().Id;
             return true;
         }
     }
